Rename some mock files
- Better name for the error mock file
- Use ts rather than tsx
